added tests for find_distance

+++ funcs.h
@@ -3,7 +3,7 @@
 // add prototypes here
 std::string encryptCaesar(std::string plaintext, int rshift);
 
-double find_distance(int v1[], int v2[], int length);
+double find_distance(double v1[], double v2[], int length);
 
 void calculate_vector(double freq_vector[], std::string encrypted_string);
